+++ src/main/java/me/tomsdevsn/hetznercloud/objects/pricing/LocationPrice.java
@@ -1,0 +1,14 @@
+package me.tomsdevsn.hetznercloud.objects.pricing;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+import lombok.Data;
+
+@Data
+public class LocationPrice {
+
+    private String location;
+    @JsonProperty("price_hourly")
+    private Price priceHourly;
+    @JsonProperty("price_monthly")
+    private Price priceMonthly;
+}